+++ frontend-desktop/src/main/java/com/pusula/desktop/dto/VehicleStockDTO.java
@@ -1,5 +1,7 @@
 package com.pusula.desktop.dto;
 
+import java.math.BigDecimal;
+
 public class VehicleStockDTO {
     private Long id;
     private Long companyId;
@@ -7,7 +9,7 @@ public class VehicleStockDTO {
     private Long inventoryId;
     private String vehicleLicensePlate;
     private String partName;
-    private Integer quantity;
+    private BigDecimal quantity;
 
     public VehicleStockDTO() {
     }
@@ -60,11 +62,11 @@ public class VehicleStockDTO {
         this.partName = partName;
     }
 
-    public Integer getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
 }
