+++ frontend-playstore/PusulaService/app/src/main/java/com/pusula/service/data/model/AdminModels.kt
@@ -36,7 +36,7 @@ data class PartProfit(
     val partName: String,
     val buyPrice: Double? = null,
     val sellPrice: Double? = null,
-    val quantitySold: Int? = null,
+    val quantitySold: Double? = null,
     val totalProfit: Double? = null,
     val marginPercent: Double? = null
 )
